tidy: Move game type to a types file
